outsourced the Firebase Login so that it can be accessible from everywhere
