+++ app/friends.html
@@ -6,6 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>MME: Projekte im SS 2022</title>
     <link rel="stylesheet" href="./resources/css/menu.css">
+    <script src="https://www.gstatic.com/firebasejs/8.6.5/firebase-app.js"></script>
+    <script src="https://www.gstatic.com/firebasejs/8.6.5/firebase-auth.js"></script>
 </head>
 
 <body>
@@ -91,10 +93,8 @@
             <div class="friends-container_elements"></div>
         </div>
     </div>
-    <a href="./index.html">
-        <button class="back-to-login-button">Back to the Login</button>
-    </a>
-    <script type="module" src="./src/search-recipe.js"></script>
+    <button id="signOut" class="back-to-login-button">Back to the Login</button>
+    <script type="module" src="./src/general-logout.js"></script>
 </body>
 
 </html>
