+++ app/index.html
@@ -36,11 +36,6 @@
                         <img src="resources/images/signup-image.png" alt="" class="login-image"> 
                         <span class="btn-text">Sign up</span>
                     </button>
-                    <a href="./search-recipe.html">
-                        <button>Go to menu</button>
-                    </a>
-                    <button id="signOut">Sign Out</button>
-                    <button id="changePassword">Change password</button>
             </div>
         </div>
     </div>
