+++ app/invitations.html
@@ -6,6 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>MME: Projekte im SS 2022</title>
     <link rel="stylesheet" href="./resources/css/menu.css">
+    <script src="https://www.gstatic.com/firebasejs/8.6.5/firebase-app.js"></script>
+    <script src="https://www.gstatic.com/firebasejs/8.6.5/firebase-auth.js"></script>
 </head>
 
 <body>
@@ -87,9 +89,8 @@
         <div class="invitations-container" id="invitation-list">
         </div>
     </div>
-    <a href="./index.html">
-        <button class="back-to-login-button">Back to the Login</button>
-    </a>
+    <button id="signOut" class="back-to-login-button">Back to the Login</button>
+
     <script type="module" src="./src/invitations/index.js"></script>
     <!-- <script type="module" src="src/index.js"></script> -->
     <template id="invitation-element">
@@ -101,6 +102,7 @@
             <button class="decline">Nein</button>
         </div>
     </template>
+    <script type="module" src="./src/general-logout.js"></script>
 </body>
 
 </html>
