+++ app/profile.html
@@ -6,6 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>MME: Projekte im SS 2022</title>
     <link rel="stylesheet" href="./resources/css/menu.css">
+    <script src="https://www.gstatic.com/firebasejs/8.6.5/firebase-app.js"></script>
+    <script src="https://www.gstatic.com/firebasejs/8.6.5/firebase-auth.js"></script>
 </head>
 
 <body>
@@ -81,9 +83,8 @@
         </div>
         <button class="go-to-settings" onclick="window.location.href='./settings.html';">Do you want to change your settings? Click here</button>
     </div>
-    <a href="./index.html">
-        <button class="back-to-login-button">Back to the Login</button>
-    </a>
+    <button id="signOut" class="back-to-login-button">Back to the Login</button>
+    <script type="module" src="./src/general-logout.js"></script>
 </body>
 
 </html>
