+++ app/resources/css/style.css
@@ -73,7 +73,7 @@ body {
 
 .login-form {
     border: 5px solid #f1f1f1;
-    height: 550px;
+    height: 400px;
     border-radius: 50px;
     padding: 60px;
     margin: 30px auto 0 auto;
@@ -126,8 +126,6 @@ body {
   }
   
   .cancelbtn {
-    width: auto;
-    padding: 10px 18px;
     background-color: #f44336;
   }
   
@@ -137,6 +135,10 @@ body {
   }
   .container {
     padding: 16px;
+  }
+
+  .buttons-content_container {
+    margin-top: 20px;
   }
 
  .login-image {
@@ -171,9 +173,14 @@ body {
     border: 1px solid gray;
   }
 
-  .popup-cancelbtn {
-    width: 150px;
-    height: 40px;
+  #signIn, #signUp {
+    width: 50%;
+    margin: 10px 25%;
+    text-align: center;
+  }
+
+  #signIn {
+    margin-top: 25px;
   }
 
   .signUpImage, .signUp-text {
@@ -224,6 +231,27 @@ body {
     height: 40px;
   }
 
+  @media screen and (max-width: 700px) {
+    .login-form {
+      padding: 10px;
+      height: 500px;
+    }
+
+    input[type=text], input[type=password] {
+      height: 60px;
+    }
+    .popup-content {
+      height: 350px;
+      width: 80%;
+    }
+
+    .signUp-text {
+      width: 50%;
+      margin: 10px auto 20px auto;
+      text-align: center;
+    }
+  }
+
   @media screen and (max-width: 650px) {
     .login-form {
       padding: 20px;
@@ -237,33 +265,18 @@ body {
     input[type=text], input[type=password] {
       height: 60px;
     }
-  }
 
-  @media screen and (max-width: 700px) {
-    .login-form {
-      padding: 10px;
-      height: 600px;
-    }
-
-    button {
-      margin: 8px 7.5%;
-      width: 85%;
-    }
-
-    input[type=text], input[type=password] {
-      height: 60px;
+    
+    #signIn, #signUp {
+      width: 90%;
+      margin-right: 5%;
+      margin-left: 5%;
     }
   }
 
   @media screen and (max-width: 420px) {
     .login-form {
       margin-top: 10px;
-      height: 600px;
-    }
-
-    button {
-      margin: 8px 7.5%;
-      width: 85%;
     }
 
     input[type=text], input[type=password] {
