+++ app/search-recipe.html
@@ -98,9 +98,7 @@
             </div>
         </div>
     </div>
-    <a href="./index.html">
-        <button class="back-to-login-button">Back to the Login</button>
-    </a>
+    <button id="signOut" class="back-to-login-button">Back to the Login</button>
     <script type="module" src="./src/search-recipe.js"></script>
     <script type="module" src="resources/js/recipeBrowser/RecipeSearch.js"></script>
 </body>
