+++ app/settings.html
@@ -84,9 +84,8 @@
         <button class="save-changes">Save changes</button>
         <button class="discard-changes">Discard changes</button>
     </div>
-    <a href="./index.html">
-        <button class="back-to-login-button">Back to the Login</button>
-    </a>
+    <button id="signOut" class="back-to-login-button">Back to the Login</button>
+    <script type="module" src="./src/general-logout.js"></script>
 </body>
 
 </html>
