+++ app/src/general-logout.js
@@ -1,0 +1,5 @@
+
+import { signOut } from "./FirebaseLogin.js";
+
+let signOutButton = document.querySelector("#signOut");
+signOutButton.addEventListener('click', signOut);
